A* Path Following Implementation
After reviewing class slides and watching the A* Explanation vide on Moodle, I tried my best to apply the knowledge I gained to the Pathfinding Task in the instructions.  There appear to be a couple issues stemming from functions not recognizing each other, so I will continue to work on it at a different date.